app/test: fix external mbuf test when assertions enabled
when RTE_ENABLE_ASSERT is defined test_mbuf application is
failing because we are trying to attach extbuf to a cloned
buffer to which external mbuf is already attached.To make
test_mbuf pass CI we have updated ol_flags. This patch fixes
the same.

Fixes: 7b295dceea07 ("test/mbuf: add unit test cases")

Signed-off-by: Rakesh Kudurumalla <rkudurumalla@marvell.com>
Signed-off-by: Nawal Kishor <nkishor@marvell.com>
Acked-by: Olivier Matz <olivier.matz@6wind.com>
Change-Id: Ib426834ce699a22ebbbd834e92664665a4f35467
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/119739
Base-Builds: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Tested-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>