net/cnxk: generalise flow operation APIs
Flow operations can be performed on cnxk ports as well as representor
ports. Since representor ports are not cnxk ports but have eswitch as
base device underneath, special handling is required to align with base
infra. Introducing a flag to generic flow APIs to discriminate if the
operation request made on normal or representor ports.

Change-Id: Ia9d07b88264c9a157bcfd8d86b609e8bb186c1f6
Signed-off-by: Harman Kalra <hkalra@marvell.com>